cloud_functions/nyse-etfs-symbol-name/requirements.txt
functions-framework==3.*
selenium
google-cloud-storage